requirements.txt
python-telegram-bot>=21.0
python-dotenv>=1.0.0
httpx[http2]>=0.27.0
Pillow>=10.0.0
flask>=3.0.0
